.vcxproj: deduplicate Debug and Release settings
By introducing new `$(BuildType)` and `$(CPU)` properties, we get to
eliminate all remaining redundancy, even in `msdos.vcxproj`. These new
properties have to be specified within a `<PropertyGroup>`, but as it
turns out, this is still a better place than `<ProjectConfiguration>`
as we get to deduplicate the platform and replace half of the
`$(BuildType)` definitions with a default value.